find enclosures directly attached to controller
All enclosures are added to m_Enclosures_Direct
so we need to check which of them really are directly
attached to controller.

At create time we can't know if enclosure is directly
attached or not because it is possible that the expanders are
daisy chained inside of one enclosure.
This would give impression enclosure is not directly attached
if we find enclosure on the second expander first.

Signed-off-by: Anna Czarnowska <anna.czarnowska@intel.com>